Workaround for ArrayIndexOutOfBoundsException Sometimes the tracker stops updating the dates. In these cases I've noticed both for myself as for other users that there is a random number on an otherwise empty line. Most recently from a friend of mine:
![image](https://cdn.discordapp.com/attachments/779907000191942716/1075452187653722235/image.png)

When you delete these lines, the tracker starts working again. I suspect that this number is the messagecount of someone or the messagecount + 1 (aka ``int newcount``). While testing my own plugins I've also noticed that clan-activity-tracker has thrown ``java.lang.ArrayIndexOutOfBoundsException``s. This makes sense, since it's likely trying to get the value for an index that does not exist (e.g. ``record.get(3)`` will result in an ``ArrayIndexOutOfBoundsException`` in the above screenshot).

I've checked the code that should increase the messagecount and that looks fine to me, so I'm not entirely sure what's the cause of these numbers. I've created this workaround so that it ignores the entry unless ``record.size() == 5``, so that should prevent any ``ArrayIndexOutOfBoundsException``s.